Move Sleeper to yolean and simplify
