+++ yolean/src/main/java/com/yahoo/yolean/UncheckedInterruptedException.java
@@ -1,5 +1,5 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.concurrent;
+package com.yahoo.yolean;
 
 /**
  * Wraps an {@link InterruptedException} with an unchecked exception.
